backend/requirements.txt
# Core Framework
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.5
pydantic-settings==2.7.1

# Database
sqlalchemy==2.0.36
alembic==1.14.0
psycopg2-binary==2.9.10
pyodbc==5.2.0

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
email-validator==2.1.0

# File Processing
python-multipart==0.0.6
pandas==2.2.3
openpyxl==3.1.5

# External Services
openai==1.58.1
stripe==11.3.0

# Utilities
python-dotenv==1.0.1
requests==2.32.3

# Caching & Sessions
redis==5.2.1

# Email
aiosmtplib==3.0.2
jinja2==3.1.5

# Testing
pytest==8.3.4
pytest-asyncio==0.24.0
httpx==0.28.1

# Monitoring & Logging
python-json-logger==3.2.1
